// src/app/core/Model/ICompany.ts
export interface ICompany {
    companyname: string;
    email: string;
    whitepapaer: string;
    website: string;
    aboutcomapny: string;
    phonenumber: string;
    address1: string;
    address2: string;
    city_id: number;
    country_id: number;
    zip_code: string;
    userid: number;
    imagename: string;
    id: number;
}